docs: update examples due to v1.0.0 of xk6-sql
